+++ footer.php
@@ -10,5 +10,28 @@
 
 <?php wp_footer(); ?>
 
+<?php if ( ! empty( get_option( 'cpt_sites_primary_menu_cta_code' ) ) ) { ?>
+  <div id="cta-modal">
+    <div class="modal-screen"></div>
+    <div class="modal card">
+
+      <?php dismiss_modal( 'cta-modal' ); ?>
+
+      <?php
+
+        $cta_code = get_option( 'cpt_sites_primary_menu_cta_code' );
+        $cta_code = wptexturize( $cta_code );
+        $cta_code = wpautop( $cta_code );
+        $cta_code = shortcode_unautop( $cta_code );
+        $cta_code = do_shortcode( $cta_code );
+
+        echo $cta_code;
+
+      ?>
+
+    </div>
+  </div>
+<?php } ?>
+
 </body>
 </html>