+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1] - 2021-12-22
+
+### Added
+-Modal call to action that can display shortcodes and embed codes.
+-Modal dismiss button (fades out modals).
+
 ## [1.0] - 2021-12-10
 
 ### Added
-- Everything
+-Everything